backend/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
django-filter==21.1.0
gunicorn==21.2.0
Pillow==10.3.0